# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeacherHub Backend API",
    "version": "1.0.0",
    "description": "Backend assessment boilerplate for Node.js + TypeScript technical tests\n\n### Quickstart\n1. Login using `POST /api/auth/login` to get a JWT.\n2. Create an API key using `POST /api/security/api-keys` (owner token required).\n3. Click **Authorize** in Swagger UI and set:\n   - `bearerAuth`: `Bearer <jwt>`\n   - `apiKeyAuth`: `<full_api_key>`\n4. Test protected endpoints (`/api/users`, `/api/schools`, `/api/teachers`, `/api/events`).\n\nSecurity model:\n- `POST /api/auth/login`: public\n- `/api/security/api-keys`: `bearerAuth`\n- Other `/api/*`: `bearerAuth + apiKeyAuth`"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "StandardResponseMeta": {
        "type": "object",
        "properties": {
          "dateTime": {
            "type": "string",
            "example": "2026-02-18 19:22:58"
          },
          "httpCode": {
            "type": "integer",
            "example": 200
          },
          "alert": {
            "type": "string",
            "enum": [
              "positive",
              "negative",
              "warning",
              "info"
            ],
            "example": "positive"
          },
          "title": {
            "type": "string",
            "example": "Correcto"
          },
          "message": {
            "type": "string",
            "example": "Consulta completada correctamente."
          }
        },
        "required": [
          "dateTime",
          "httpCode",
          "alert",
          "title",
          "message"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "UserMembership": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid",
            "example": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa"
          },
          "schoolName": {
            "type": "string",
            "example": "TeacherHub Academy"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "teacher",
              "student"
            ]
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "schoolId",
          "schoolName",
          "role",
          "isDefault"
        ]
      },
      "UserMembershipInput": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid",
            "example": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa"
          },
          "schoolName": {
            "type": "string",
            "example": "TeacherHub Academy"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "teacher",
              "student"
            ]
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "schoolId",
          "schoolName",
          "role"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "owner",
              "user"
            ],
            "example": "user"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "example": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "teacher",
              "student"
            ]
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembershipInput"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "globalRole",
          "tenantId",
          "role",
          "memberships",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 4
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "owner",
              "user"
            ],
            "default": "user"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembershipInput"
            }
          }
        },
        "required": [
          "email",
          "password",
          "memberships"
        ]
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 4
          },
          "globalRole": {
            "type": "string",
            "enum": [
              "owner",
              "user"
            ]
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembership"
            }
          }
        }
      },
      "UsersListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserProfileResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "UserStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserProfileResponse"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "UsersListStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UsersListResponse"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "UserDeleteStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "id"
                ]
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user with tenant memberships",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiUsers"
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "globalRole",
            "schema": {
              "type": "string",
              "enum": [
                "owner",
                "user"
              ]
            }
          },
          {
            "in": "query",
            "name": "schoolId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "email",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiUsers"
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiUsersId"
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user and memberships",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "putApiUsersId"
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeleteStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteApiUsersId"
      }
    },
...
  }
}